# CryptoAlg-682: Improve AES-GCM on ARMv8