--- a/crypto/fipsmodule/modes/asm/ghashv8-armx.pl
+++ b/crypto/fipsmodule/modes/asm/ghashv8-armx.pl
@@ -32,13 +32,13 @@
 #
 # Current performance in cycles per processed byte:
 #
-#		   64-bit PMULL	 32-bit PMULL	32-bit NEON(*)
-# Apple A7	    0.58		0.92		5.62
-# Cortex-A53	0.85		1.01		8.39
-# Cortex-A57	0.73		1.17		7.61
-# Denver	    0.51		0.65		6.02
-# Mongoose	    0.65		1.10		8.06
-# Kryo		    0.76		1.16		8.00
+#           64-bit PMULL  32-bit PMULL  32-bit NEON(*)
+# Apple A7      0.58         0.92          5.62
+# Cortex-A53    0.85         1.01          8.39
+# Cortex-A57    0.73         1.17          7.61
+# Denver        0.51         0.65          6.02
+# Mongoose      0.65         1.10          8.06
+# Kryo          0.76         1.16          8.00
 #
 # (*)	presented for reference/comparison purposes;