### Issues:
Resolves CryptoAlg-682

### Description of changes: 
AWS-LC lags OpenSSL by 14-18% in the performance of AES-GCM-128/256 on Armv8, while it exceeds it on x86_64.
This PR merges in the changes in [OpenSSL's ghashv8-armx.pl](https://github.com/openssl/openssl/blob/90cebd1b216e0a160fcfd8e0eddca47dad47c183/crypto/modes/asm/ghashv8-armx.pl) as in [OpenSSL 1.1.1k](https://github.com/openssl/openssl/releases/tag/OpenSSL_1_1_1k).
These changes extend Gueron's method from 2x-aggregate factor to 4x-aggregate (see slide 22 in [Gueron's slides](https://crypto.stanford.edu/RealWorldCrypto/slides/gueron.pdf)) and matches the implementation in ghash-x86_64.pl

### Call-outs:
N/A

### Testing:
Speed tests were run on ARMv8, 2.5GHz:
```
Before:
Did 204000 AES-128-GCM (8192 bytes) seal operations in 1004479us (203090.4 ops/sec): 1663.7 MB/s
Did 103000 AES-128-GCM (16384 bytes) seal operations in 1002904us (102701.8 ops/sec): 1682.7 MB/s
Did 173000 AES-256-GCM (8192 bytes) seal operations in 1000072us (172987.5 ops/sec): 1417.1 MB/s
Did 88000 AES-256-GCM (16384 bytes) seal operations in 1008040us (87298.1 ops/sec): 1430.3 MB/s

After:
Did 252000 AES-128-GCM (8192 bytes) seal operations in 1001478us (251628.1 ops/sec): 2061.3 MB/s
Did 128000 AES-128-GCM (16384 bytes) seal operations in 1003033us (127612.9 ops/sec): 2090.8 MB/s
Did 208000 AES-256-GCM (8192 bytes) seal operations in 1004083us (207154.2 ops/sec): 1697.0 MB/s
Did 105000 AES-256-GCM (16384 bytes) seal operations in 1002262us (104763.0 ops/sec): 1716.4 MB/s
```

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.